requirements.txt
requests
pycryptodomex
python-telegram
python-brightid